+++ src/Service/Change/Command/TransferChangesCommand.php
@@ -64,6 +64,7 @@ class TransferChangesCommand
                 false,
                 true,
                 false,
+                false,
                 false
             )
         );